// source/modules/team/team.actions-enum.ts
export enum TeamActionTypes {
	GET_TEAMS = 'GET_TEAMS',
	GET_TEAMS_SUCCESS = 'GET_TEAMS_SUCCESS',
	GET_TEAMS_FAILURE = 'GET_TEAMS_FAILURE',
	GET_TASKS = 'GET_TASKS',
	GET_TASKS_SUCCESS = 'GET_TASKS_SUCCESS',
	GET_TASKS_FAILURE = 'GET_TASKS_FAILURE',
	SET_TASK = 'SET_TASK',
	SET_TASK_SUCCESS = 'SET_TASK_SUCCESS',
	SET_TASK_FAILURE = 'SET_TASK_FAILURE',
	DELETE_TASK = 'DELETE_TASK',
	DELETE_TASK_SUCCESS = 'DELETE_TASK_SUCCESS',
	DELETE_TASK_FAILURE = 'DELETE_TASK_FAILURE',
	ON_TASKS_SNAPSHOT_SUCCESS = 'ON_TASKS_SNAPSHOT_SUCCESS',
	ON_TASKS_SNAPSHOT_FAILURE = 'ON_TASKS_SNAPSHOT_FAILURE',
}